chore: only one spark session
