+++ source/geh_calculated_measurements/tests/conftest.py
@@ -1,0 +1,36 @@
+### This file contains the fixtures that are used in the tests. ###
+from typing import Generator
+
+import pytest
+from delta import configure_spark_with_delta_pip
+from pyspark.sql import SparkSession
+
+
+@pytest.fixture(scope="module", autouse=True)
+def clear_cache(spark: SparkSession) -> Generator[None, None, None]:
+    """
+    Clear the cache after each test module to avoid memory issues.
+    """
+    yield
+    spark.catalog.clearCache()
+
+
+@pytest.fixture(scope="session")
+def spark() -> Generator[SparkSession, None, None]:
+    """
+    Create a Spark session with Delta Lake enabled.
+    """
+    session = (
+        SparkSession.builder.appName("geh_calculated_measurements")  # # type: ignore
+        .config("spark.sql.extensions", "io.delta.sql.DeltaSparkSessionExtension")
+        .config(
+            "spark.sql.catalog.spark_catalog",
+            "org.apache.spark.sql.delta.catalog.DeltaCatalog",
+        )
+        # Enable Hive support for persistence across test sessions
+        .config("spark.sql.catalogImplementation", "hive")
+        .enableHiveSupport()
+    )
+    session = configure_spark_with_delta_pip(session).getOrCreate()
+    yield session
+    session.stop()
